+++ common/src/main/java/earth/terrarium/argonauts/common/compat/heracles/ArgonautsTeamProvider.java
@@ -3,7 +3,6 @@ package earth.terrarium.argonauts.common.compat.heracles;
 import com.mojang.authlib.GameProfile;
 import earth.terrarium.argonauts.api.guild.Guild;
 import earth.terrarium.argonauts.api.guild.GuildApi;
-import earth.terrarium.argonauts.common.handlers.base.members.MemberState;
 import earth.terrarium.argonauts.common.handlers.guild.members.GuildMember;
 import earth.terrarium.heracles.api.teams.TeamProvider;
 import net.minecraft.server.level.ServerLevel;
@@ -29,7 +28,7 @@ public class ArgonautsTeamProvider implements TeamProvider {
         return Stream.of(guild.members()
             .allMembers()
             .stream()
-            .filter(member -> member.getState() != MemberState.INVITED)
+            .filter(member -> member.getState().isPermanentMember())
             .map(GuildMember::profile)
             .map(GameProfile::getId)
             .filter(uuid -> !uuid.equals(player))