+++ common/src/main/java/earth/terrarium/argonauts/common/handlers/base/members/MemberState.java
@@ -3,9 +3,15 @@ package earth.terrarium.argonauts.common.handlers.base.members;
 public enum MemberState {
     OWNER,
     MEMBER,
-    INVITED;
+    INVITED,
+    ALLIED,
+    ;
 
     public boolean isLeader() {
         return this == OWNER;
+    }
+
+    public boolean isPermanentMember() {
+        return this == MEMBER || isLeader();
     }
 }